backend/requirements.txt
flask==3.0.2
flask_cors==4.0.0
PyPDF2==3.0.1
spacy==3.7.4
python-dateutil==2.9.0